+++ src/pages/login.tsx
@@ -1,7 +1,8 @@
 import styled from '@emotion/styled';
 import ImageLogo from 'public/images/image-logo.svg';
 import LetterLogo from 'public/images/letter-logo.svg';
-import GoogleLogo from 'public/images/google-logo.svg';
+
+import GoogleLoginButton from '@/components/login/GoogleLoginButton';
 
 export default function LoginPage() {
   return (
@@ -9,10 +10,7 @@ export default function LoginPage() {
       <StyledImageLogo />
       <StyledLetterLogo />
       <Slogan>What’s your Hype Music?</Slogan>
-      <GoogleLoginButton>
-        <StyledGoogleLogo />
-        Google로 시작하기
-      </GoogleLoginButton>
+      <GoogleLoginButton />
       <Footer>© 6Jeans. All rights reversed.</Footer>
     </Container>
   );
@@ -42,27 +40,6 @@ const Slogan = styled.div`
   line-height: 3rem;
   margin-top: 5.2rem;
   margin-bottom: auto;
-`;
-
-const GoogleLoginButton = styled.button`
-  position: relative;
-  background-color: #ffffff;
-  border: 0.4px solid #b7b7b7;
-  border-radius: 10px;
-  width: calc(100% - 6.8rem);
-  padding: 1.5rem 0;
-  font-family: 'Poppins';
-  font-weight: 600;
-  font-size: 15px;
-  line-height: 22px;
-  color: #848484;
-  margin: 0 3.4rem;
-`;
-
-const StyledGoogleLogo = styled(GoogleLogo)`
-  position: absolute;
-  left: 2.2rem;
-  top: 1.7rem;
 `;
 
 const Footer = styled.footer`
